+++ hnix-store-nar/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Switch from `cryptonite` to `crypton`
+* Preserve non-UTF-8 filename and symlink target bytes when streaming NARs
 
 # [0.1.1.0](https://github.com/haskell-nix/hnix-store/compare/nar-0.1.0.0...nar-0.1.1.0)  2024-10-09